examples/vision_integration.py
# CraftX-ML Vision Tool Integration Example

"""
Example of how CraftX-ML orchestrates external vision tools
instead of relying on built-in multimodal capabilities.
"""

import cv2
import pytesseract
from PIL import Image
import base64
import hashlib


class VisionToolOrchestrator:
    """External vision processing for CraftX-ML sovereignty."""

    def __init__(self):
        self.tools = {
            'ocr': self.extract_text,
            'analyze': self.analyze_image,
            'screenshot': self.process_screenshot,
            'logo_convert': self.convert_logo
        }

    def extract_text(self, image_path):
        """OCR text extraction with attestation logging."""
        image = Image.open(image_path)
        text = pytesseract.image_to_string(image)

        # Attestation logging
        hash_val = hashlib.sha256(text.encode()).hexdigest()

        return {
            'extracted_text': text,
            'attestation_hash': hash_val,
            'tool': 'ocr',
            'status': 'success'
        }

    def analyze_image(self, image_path):
        """Structural image analysis."""
        image = cv2.imread(image_path)

        # Basic analysis
        height, width, channels = image.shape

        return {
            'dimensions': f"{width}x{height}",
            'channels': channels,
            'analysis': 'structural_complete',
            'tool': 'analyze'
        }

    def process_screenshot(self, image_path):
        """Screenshot-specific processing."""
        # Combine OCR + structural analysis for UI elements
        ocr_result = self.extract_text(image_path)
        structure = self.analyze_image(image_path)

        return {
            'ui_text': ocr_result['extracted_text'],
            'ui_structure': structure,
            'tool': 'screenshot'
        }

# Tool call protocol for CraftX-ML


def vision_tool_call(tool_name, image_path):
    """
    Protocol for CraftX-ML to invoke external vision tools.
    Maintains sovereignty and auditability.
    """
    orchestrator = VisionToolOrchestrator()

    if tool_name in orchestrator.tools:
        result = orchestrator.tools[tool_name](image_path)

        # Log for attestation
        print(f"[CRAFTX-ML] Vision tool '{tool_name}' executed")
        print(
            f"[ATTESTATION] Result hash: {result.get('attestation_hash', 'N/A')}")

        return result
    else:
        return {'error': f'Unknown vision tool: {tool_name}'}


if __name__ == "__main__":
    # Example usage
    print("CraftX-ML Vision Tool Integration Ready")
    print("Available tools:", list(VisionToolOrchestrator().tools.keys()))
