requirements.txt
streamlit>=1.28.0
pytest>=7.0.0
requests>=2.28.0
Pillow>=8.0.0
cryptography>=3.4.8
pyjwt>=2.4.0
oauthlib>=3.2.0
google-auth>=2.0.0
google-auth-oauthlib>=0.5.0
msal>=1.20.0
authlib>=1.0.0
boto3>=1.26.0
azure-storage-blob>=12.0.0
dropbox>=11.0.0
pandas>=1.5.0
plotly>=5.0.0
python-dateutil>=2.8.0
dnspython>=2.7.0
# Optional provider clients (only used if available)
ollama>=0.3.0
openai>=1.40.0
